pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.0",
    "websockets>=14.1",
]
